+++ apps/OboeTester/app/src/main/cpp/TestRapidCycle.cpp
@@ -18,7 +18,7 @@
 #include <aaudio/AAudioExtensions.h>
 
 #include "common/OboeDebug.h"
-#include "common/AudioClock.h"
+#include "oboe/AudioClock.h"
 #include "TestRapidCycle.h"
 
 using namespace oboe;
