+++ include/oboe/Oboe.h
@@ -36,5 +36,6 @@
 #include "oboe/FifoBuffer.h"
 #include "oboe/OboeExtensions.h"
 #include "oboe/FullDuplexStream.h"
+#include "oboe/AudioClock.h"
 
 #endif //OBOE_OBOE_H
