+++ src/aaudio/AudioStreamAAudio.cpp
@@ -20,8 +20,8 @@
 
 #include "aaudio/AAudioLoader.h"
 #include "aaudio/AudioStreamAAudio.h"
-#include "common/AudioClock.h"
 #include "common/OboeDebug.h"
+#include "oboe/AudioClock.h"
 #include "oboe/Utilities.h"
 #include "AAudioExtensions.h"
