+++ src/common/AdpfWrapper.cpp
@@ -18,8 +18,8 @@
 #include <stdint.h>
 #include <sys/types.h>
 
+#include "oboe/AudioClock.h"
 #include "AdpfWrapper.h"
-#include "AudioClock.h"
 #include "OboeDebug.h"
 
 typedef APerformanceHintManager* (*APH_getManager)();
