+++ src/opensles/AudioOutputStreamOpenSLES.cpp
@@ -16,9 +16,8 @@
 
 #include <cassert>
 
-#include <common/AudioClock.h>
-
 #include "common/OboeDebug.h"
+#include "oboe/AudioClock.h"
 #include "oboe/AudioStreamBuilder.h"
 #include "AudioOutputStreamOpenSLES.h"
 #include "AudioStreamOpenSLES.h"
